Move containersource OIDC feature to test/auth
