+++ test/auth/oidc_test.go
@@ -32,7 +32,6 @@ import (
 	"knative.dev/eventing/test/auth/features/oidc"
 	brokerfeatures "knative.dev/eventing/test/rekt/features/broker"
 	"knative.dev/eventing/test/rekt/features/channel"
-	"knative.dev/eventing/test/rekt/features/containersource"
 	parallelfeatures "knative.dev/eventing/test/rekt/features/parallel"
 	sequencefeatures "knative.dev/eventing/test/rekt/features/sequence"
 	"knative.dev/eventing/test/rekt/resources/broker"
@@ -157,7 +156,7 @@ func TestContainerSourceSendsEventsWithOIDCSupport(t *testing.T) {
 		environment.Managed(t),
 	)
 
-	env.Test(ctx, t, containersource.SendsEventsWithSinkRefOIDC())
+	env.Test(ctx, t, oidc.SendsEventsWithSinkRefOIDC())
 }
 
 func TestSequenceSendsEventsWithOIDCSupport(t *testing.T) {
